% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wetlands_syntax.R
\docType{data}
\name{grasslands}
\alias{grasslands}
\title{Grasslands from Temuco, Chile}
\format{
A \linkS4class{vegtable} object.
}
\usage{
grasslands
}
\description{
Grasslands from Temuco, Chile
}
\details{
Describe.
}
\references{
\bold{San Martín C et al. (1998).}
La vegetación de lagunas primaverales en las cercanías de Temuco
(Cautín, Chile).
\emph{Acta Botanica Malacitana} 23: 99--120.
}
\keyword{datasets}
