% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mullenders_t2.R
\docType{data}
\name{mullenders_t2}
\alias{mullenders_t2}
\title{Association of Siegesbeckia and Cynoglossum}
\format{
An object of class \linkS4class{data.frame}.
}
\usage{
mullenders_t2
}
\description{
Association of Siegesbeckia and Cynoglossum
}
\details{
This table contain a cross table of species by plots including the cover of
species in the \bold{Braun-Blanquet} scale.
The first column includes the species names as published in the original
source (see references).
The following columns are the plots identified by their entries in the data
base \href{http://www.givd.info/ID/AF-00-006}{SWEA-Dataveg}.
}
\examples{
data(mullenders_t2)
head(mullenders_t2)

}
\references{
\bold{Alvarez M et al. (2012).} SWEA-Dataveg -- Vegetation of small wetlands
in East Africa. \emph{Biodiversity and Ecology} 4: 294--295.

\bold{Mullenders W (1953).} Contribution a l'étude des groupements végétaux
de la contrée de Goma-Kisenyi (Kivu-Ruanda). \emph{Vegetatio} 4: 73--83.
url: \url{https://www.jstor.org/stable/20034547?seq=1}
}
\keyword{datasets}
