% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wetlands_syntax.R
\docType{data}
\name{wetlands_syntax}
\alias{wetlands_syntax}
\alias{wetland_syntax}
\title{Syntaxonomic List of East African Semi-Aquatic Vegetation}
\format{
A list of three data frames.
}
\usage{
wetland_syntax
}
\description{
Syntaxonomic List of East African Semi-Aquatic Vegetation
}
\details{
Describe.
}
\references{
\bold{Alvarez M et al. (2012).} SWEA-Dataveg -- Vegetation of small wetlands
in East Africa. \emph{Biodiversity and Ecology} 4: 294--295.
}
\keyword{datasets}
